LPS-5942 Move
