+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/instance/lifecycle/CheckJournalContentSearchPortalInstanceLifecycleListener.java
@@ -12,11 +12,12 @@
  * details.
  */
 
-package com.liferay.journal.lifecycle;
+package com.liferay.journal.instance.lifecycle;
 
 import com.liferay.journal.configuration.JournalServiceConfigurationValues;
 import com.liferay.journal.service.JournalContentSearchLocalService;
 import com.liferay.portal.instance.lifecycle.BasePortalInstanceLifecycleListener;
+import com.liferay.portal.instance.lifecycle.PortalInstanceLifecycleListener;
 import com.liferay.portal.kernel.module.framework.ModuleServiceLifecycle;
 import com.liferay.portal.model.Company;
 
@@ -28,7 +29,7 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(immediate = true, service = PortalInstanceLifecycleListener.class)
 public class CheckJournalContentSearchPortalInstanceLifecycleListener
-	extends PortalInstanceLifecycleListener {
+	extends BasePortalInstanceLifecycleListener {
 
 	@Override
 	public void portalInstanceRegistered(Company company) throws Exception {
