+++ modules/apps/wiki/wiki-layout-prototype/src/main/java/com/liferay/wiki/layout/prototype/instance/lifecycle/AddLayoutPrototypePortalInstanceLifecycleListener.java
@@ -12,9 +12,10 @@
  * details.
  */
 
-package com.liferay.wiki.layout.prototype.lifecycle;
+package com.liferay.wiki.layout.prototype.instance.lifecycle;
 
 import com.liferay.portal.instance.lifecycle.BasePortalInstanceLifecycleListener;
+import com.liferay.portal.instance.lifecycle.PortalInstanceLifecycleListener;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.module.framework.ModuleServiceLifecycle;
 import com.liferay.portal.model.Company;
@@ -37,7 +38,7 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(immediate = true, service = PortalInstanceLifecycleListener.class)
 public class AddLayoutPrototypePortalInstanceLifecycleListener
-	extends PortalInstanceLifecycleListener {
+	extends BasePortalInstanceLifecycleListener {
 
 	@Override
 	public void portalInstanceRegistered(Company company) throws Exception {
