+++ modules/portal/portal-instance-lifecycle/src/main/java/com/liferay/portal/instance/lifecycle/internal/PortalInstanceLifecycleListenerManager.java
@@ -14,7 +14,7 @@
 
 package com.liferay.portal.instance.lifecycle.internal;
 
-import com.liferay.portal.instance.lifecycle.BasePortalInstanceLifecycleListener;
+import com.liferay.portal.instance.lifecycle.PortalInstanceLifecycleListener;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.module.framework.ModuleServiceLifecycle;
